backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
python-dotenv==1.0.0
httpx==0.26.0
openai==1.12.0
pyyaml==6.0.1
requests==2.31.0
